Cleaned up code in List component and organized file structure
