--- bio-bid/src/components/company-index/test-config.js
@@ -1,2 +1,0 @@
-import '@testing-library/jest-dom/extend-expect';
-import 'mutationobserver-shim';
